Authorization basics #22

+++ ideas/app/Http/Controllers/IdeaController.php
@@ -8,40 +8,59 @@ use Illuminate\Http\Request;
 class IdeaController extends Controller
 {
 
-    public function show(Idea $idea){
+    public function show(Idea $idea)
+    {
         return view('ideas.show', compact('idea'));
     }
 
-    public function store(){
+    public function store()
+    {
         $validated = request()->validate([
-                'content' => 'required|min:3|max:240',
-            ]);
+            'content' => 'required|min:3|max:240',
+        ]);
+
+        $validated['user_id'] = auth()->id();
 
         $idea = Idea::create($validated);
 
-        return redirect()->route('dashboard')->with('success','Idea created successfully');
+        return redirect()->route('dashboard')->with('success', 'Idea created successfully');
 
     }
 
-    public function destroy(Idea $idea){
+    public function destroy(Idea $idea)
+    {
+        if (auth()->id() !== $idea->user_id) {
+            abort(404);
+        };
+
         $idea->delete();
-        return redirect()->route('dashboard')->with('success','Idea deleted successfully');
+        return redirect()->route('dashboard')->with('success', 'Idea deleted successfully');
     }
 
-    public function edit(Idea $idea) {
+    public function edit(Idea $idea)
+    {
+        if (auth()->id() !== $idea->user_id) {
+            abort(404);
+        };
+
         $editing = true;
         return view('ideas.show', compact('idea', 'editing'));
     }
 
-    public function update(Idea $idea) {
+    public function update(Idea $idea)
+    {
+        if (auth()->id() !== $idea->user_id) {
+            abort(404);
+        };
+
         $validated = request()->validate([
-                'content' => 'required|min:3|max:240',
-            ]
+            'content' => 'required|min:3|max:240',
+        ]
         );
 
         $idea->update($validated);
 
-        return redirect()->route('ideas.show', $idea->id)->with('success','Idea updated successfully');
+        return redirect()->route('ideas.show', $idea->id)->with('success', 'Idea updated successfully');
     }
 
 }
